+++ src/GOODFFrame.cpp
@@ -9,11 +9,11 @@
 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
 
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  * You can contact the author on larspalo(at)yahoo DOT se
  */
@@ -32,6 +32,7 @@
 #include "CopyElementAttributesDialog.h"
 #include "CmbDialog.h"
 #include "DefaultPathsDialog.h"
+#include "StopRankImportDialog.h"
 #include <vector>
 #include <algorithm>
 
@@ -44,6 +45,7 @@ BEGIN_EVENT_TABLE(GOODFFrame, wxFrame)
 	EVT_MENU(ID_NEW_ORGAN, GOODFFrame::OnNewOrgan)
 	EVT_MENU(ID_READ_ORGAN, GOODFFrame::OnReadOrganFile)
 	EVT_MENU(ID_IMPORT_VOICING_DATA, GOODFFrame::OnImportCMB)
+	EVT_MENU(ID_IMPORT_STOP_RANK, GOODFFrame::OnImportStopRank)
 	EVT_MENU(ID_GLOBAL_SHOW_TOOLTIPS_OPTION, GOODFFrame::OnEnableTooltipsMenu)
 	EVT_MENU(ID_CLEAR_HISTORY, GOODFFrame::OnClearHistory)
 	EVT_MENU(ID_DEFAULT_PATHS_MENU, GOODFFrame::OnDefaultPathMenuChoice)
@@ -93,6 +95,7 @@ GOODFFrame::GOODFFrame(const wxString& title) : wxFrame(NULL, wxID_ANY, title) {
 
 	// Add tools menu items
 	m_toolsMenu->Append(ID_IMPORT_VOICING_DATA, wxT("Import .cmb\tCtrl+I"), wxT("Import voicing data from a .cmb (settings) file"));
+	m_toolsMenu->Append(ID_IMPORT_STOP_RANK, wxT("Import Stops/Ranks\tCtrl+R"), wxT("Import stops/ranks from another (working) .organ file"));
 	m_toolsMenu->AppendCheckItem(ID_GLOBAL_SHOW_TOOLTIPS_OPTION, wxT("Enable Tooltips"), wxT("Enable tooltips for certain controls"));
 	m_toolsMenu->Check(ID_GLOBAL_SHOW_TOOLTIPS_OPTION, false);
 	m_toolsMenu->Append(ID_CLEAR_HISTORY, wxT("Clear File History"), wxT("Remove all the entries in the recent file history"));
@@ -2552,6 +2555,34 @@ void GOODFFrame::AddStopItemToTree() {
 	m_organ->setModified(true);
 }
 
+void GOODFFrame::AddStopItemForManual(wxString stopName, unsigned manIndex) {
+	if (manIndex < m_organ->getNumberOfManuals() && manIndex < m_organTreeCtrl->GetChildrenCount(tree_manuals)) {
+		unsigned numChildrens = m_organTreeCtrl->GetChildrenCount(tree_manuals);
+		wxTreeItemIdValue cookie;
+		wxTreeItemId theManual;
+		for (unsigned i = 0; i < numChildrens; i++) {
+
+			if (i == 0)
+				theManual = m_organTreeCtrl->GetFirstChild(tree_manuals, cookie);
+			else
+				theManual = m_organTreeCtrl->GetNextChild(tree_manuals, cookie);
+			if (i == manIndex)
+				break;
+		}
+
+		wxTreeItemId divisionalChild = m_organTreeCtrl->GetLastChild(theManual);
+		wxTreeItemId couplerChild = m_organTreeCtrl->GetPrevSibling(divisionalChild);
+		wxTreeItemId stopChild = m_organTreeCtrl->GetPrevSibling(couplerChild);
+		m_organTreeCtrl->AppendItem(stopChild, stopName);
+		m_organ->setModified(true);
+	}
+}
+
+void GOODFFrame::AddRankItemToTree(wxString rankName) {
+	m_organTreeCtrl->AppendItem(tree_ranks, rankName);
+	m_organ->setModified(true);
+}
+
 void GOODFFrame::SelectStopItemInTree(int nbrAdded) {
 	wxTreeItemId selectedManual = m_organTreeCtrl->GetSelection();
 	wxTreeItemId divisionalChild = m_organTreeCtrl->GetLastChild(selectedManual);
@@ -2599,6 +2630,8 @@ void GOODFFrame::OnAddNewWindchestgroup(wxCommandEvent& WXUNUSED(event)) {
 		wxTreeItemId firstAdded;
 		for (int i = 0; i < nbrToAdd; i++) {
 			Windchestgroup newWindchest;
+			newWindchest.setIsPercussive(m_organ->getIsPercussive());
+			newWindchest.setHasIndependentRelease(m_organ->getHasIndependentRelease());
 			m_organ->addWindchestgroup(newWindchest);
 			if (i == 0)
 				firstAdded = m_organTreeCtrl->AppendItem(tree_windchestgrps, newWindchest.getName());
@@ -2639,6 +2672,8 @@ void GOODFFrame::OnAddNewRank(wxCommandEvent& WXUNUSED(event)) {
 		wxTreeItemId firstAdded;
 		for (int i = 0; i < nbrToAdd; i++) {
 			Rank newRank;
+			newRank.setPercussive(m_organ->getIsPercussive());
+			newRank.setIndependentRelease(m_organ->getHasIndependentRelease());
 			m_organ->addRank(newRank);
 			if (i == 0)
 				firstAdded = m_organTreeCtrl->AppendItem(tree_ranks, newRank.getName());
@@ -2777,8 +2812,11 @@ void GOODFFrame::OnImportCMB(wxCommandEvent& WXUNUSED(event)) {
 	if (importCmb.ShowModal() == wxID_OK) {
 		CMB_ORGAN *imported = importCmb.GetCmbOrgan();
 
-		if (m_organ->getNumberOfRanks() != imported->cmbRanks.size() || m_organ->getNumberOfStops() != imported->cmbStops.size()) {
-			wxMessageDialog dlg(this, wxT("The number of ranks/stops in the imported cmb doesn't match current organ! Do you want to import selected data to what is possible anyway?"), wxT("Are you really sure?"), wxYES_NO|wxCENTRE|wxICON_EXCLAMATION);
+		if (m_organ->getNumberOfRanks() != imported->cmbRanks.size() ||
+			m_organ->getNumberOfStops() != imported->cmbStops.size() ||
+			m_organ->getNumberOfWindchestgroups() != imported->cmbWindchests.size()
+		) {
+			wxMessageDialog dlg(this, wxT("The number of windchests/ranks/stops in the imported cmb doesn't match current organ! Do you want to import selected data to what is possible anyway?"), wxT("Are you really sure?"), wxYES_NO|wxCENTRE|wxICON_EXCLAMATION);
 			if (dlg.ShowModal() != wxID_YES) {
 				return;
 			}
@@ -2818,6 +2856,50 @@ void GOODFFrame::OnImportCMB(wxCommandEvent& WXUNUSED(event)) {
 		}
 		if (importCmb.GetImportTrackerDelay()) {
 			m_organ->setTrackerDelay(imported->attributes.trackerDelay);
+		}
+
+		unsigned windchestIdx = 0;
+		for (CMB_ELEMENT &w : imported->cmbWindchests) {
+			if (m_organ->getNumberOfWindchestgroups() > windchestIdx) {
+				Windchestgroup *windchest = m_organ->getOrganWindchestgroupAt(windchestIdx);
+
+				if (importCmb.GetImportAmplitude())
+					windchest->setAmplitudeLevel(w.amplitude);
+				if (importCmb.GetImportGain())
+					windchest->setGain(w.gain);
+				if (importCmb.GetImportPitchTuning()) {
+					float pitchT = windchest->getPitchTuning() + w.pitchTuning;
+					if (pitchT >= -1800 && pitchT <= 1800) {
+						windchest->setPitchTuning(pitchT);
+					} else {
+						wxMessageDialog pTerr(this, wxT("The PitchTuning value would be out of bounds! Do you want to set it to max allowed value?"), wxT("Value is out of bounds!"), wxYES_NO|wxCENTRE|wxICON_EXCLAMATION);
+						if (pTerr.ShowModal() == wxID_YES) {
+							if (pitchT > 1800)
+								windchest->setPitchTuning(1800);
+							else if (pitchT < -1800)
+								windchest->setPitchTuning(-1800);
+						}
+					}
+				}
+				if (importCmb.GetImportPitchCorrection()) {
+					float pitchC = windchest->getPitchCorrection() + w.pitchCorrection;
+					if (pitchC >= -1800 && pitchC <= 1800) {
+						windchest->setPitchCorrection(pitchC);
+					} else {
+						wxMessageDialog pCerr(this, wxT("The PitchCorrection value would be out of bounds! Do you want to set it to max allowed value?"), wxT("Value is out of bounds!"), wxYES_NO|wxCENTRE|wxICON_EXCLAMATION);
+						if (pCerr.ShowModal() == wxID_YES) {
+							if (pitchC > 1800)
+								windchest->setPitchCorrection(1800);
+							else if (pitchC < -1800)
+								windchest->setPitchCorrection(-1800);
+						}
+					}
+				}
+				if (importCmb.GetImportTrackerDelay()) {
+					windchest->setTrackerDelay(w.trackerDelay);
+				}
+			}
+			windchestIdx++;
 		}
 
 		unsigned stopIdx = 0;
@@ -3031,6 +3113,42 @@ void GOODFFrame::OnDefaultPathMenuChoice(wxCommandEvent& WXUNUSED(event)) {
 		m_defaultOrganDirectory = defaultPaths.GetSelectedOrganDirectory();
 		m_defaultCmbDirectory = defaultPaths.GetSelectedCmbDirectory();
 	}
+}
+
+void GOODFFrame::OnImportStopRank(wxCommandEvent& WXUNUSED(event)) {
+	wxString organFilePath;
+	wxString defaultPath = m_defaultOrganDirectory;
+	if (defaultPath == wxEmptyString)
+		defaultPath = wxStandardPaths::Get().GetDocumentsDir();
+
+	wxFileDialog fileDialog(
+		this,
+		wxT("Select .organ file to import stops/ranks from"),
+		defaultPath,
+		"",
+		"GrandOrgue ODF files (*.organ)|*.organ;*.ORGAN",
+		wxFD_OPEN|wxFD_FILE_MUST_EXIST
+	);
+
+	if (fileDialog.ShowModal() != wxID_OK) {
+		return;
+	}
+
+	organFilePath = fileDialog.GetPath();
+
+	Organ *sourceOrgan = new Organ();
+
+	OrganFileParser parser(organFilePath, sourceOrgan);
+	if (parser.isOrganReady()) {
+		StopRankImportDialog importDialog(sourceOrgan, m_organ, this);
+		importDialog.ShowModal();
+
+	} else {
+		wxMessageDialog msg(this, wxT("The selected .organ file could not be parsed for importing any stops/ranks!"), wxT("Failure to parse .organ file"), wxOK|wxCENTRE|wxICON_ERROR);
+		msg.ShowModal();
+	}
+
+	delete sourceOrgan;
 }
 
 void GOODFFrame::SetupOrganMainPanel() {