+++ src/GOODFFrame.h
@@ -9,11 +9,11 @@
 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
 
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  * You can contact the author on larspalo(at)yahoo DOT se
  */
@@ -66,6 +66,8 @@ public:
 	void RemoveCurrentItemFromOrgan();
 
 	void AddStopItemToTree();
+	void AddStopItemForManual(wxString stopName, unsigned manIndex);
+	void AddRankItemToTree(wxString rankName);
 	void SelectStopItemInTree(int nbrAdded);
 	void AddCouplerItemToTree();
 	void AddDivisionalItemToTree();
@@ -179,6 +181,7 @@ private:
 	void OnClearHistory(wxCommandEvent& event);
 	void OnSizeChange(wxSizeEvent& event);
 	void OnDefaultPathMenuChoice(wxCommandEvent& event);
+	void OnImportStopRank(wxCommandEvent& event);
 
 	void SetupOrganMainPanel();
 	void removeAllItemsFromTree();