+++ src/Rank.h
@@ -9,11 +9,11 @@
  *
  * GOODF is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with GOODF.  If not, see <https://www.gnu.org/licenses/>.
+ * along with GOODF. If not, see <https://www.gnu.org/licenses/>.
  *
  * You can contact the author on larspalo(at)yahoo.se
  */
@@ -28,14 +28,17 @@
 #include <wx/dir.h>
 #include <wx/fileconf.h>
 
+class Organ;
+
 class Rank {
 public:
 	Rank();
+	Rank(const Rank& r);
 	~Rank();
 
 	void write(wxTextFile *outFile);
 	void writeFromStop(wxTextFile *outFile);
-	void read(wxFileConfig *cfg);
+	void read(wxFileConfig *cfg, Organ *readOrgan);
 
 	bool doesAcceptsRetuning() const;
 	void setAcceptsRetuning(bool acceptsRetuning);
@@ -57,6 +60,8 @@ public:
 	void setNumberOfLogicalPipes(int numberOfLogicalPipes);
 	bool isPercussive() const;
 	void setPercussive(bool percussive);
+	bool isIndependentRelease();
+	void setIndependentRelease(bool independent);
 	float getPitchCorrection() const;
 	void setPitchCorrection(float pitchCorrection);
 	float getPitchTuning() const;
@@ -135,6 +140,7 @@ protected:
 	float pitchCorrection;
 	Windchestgroup* windchest;
 	bool percussive;
+	bool hasIndependentRelease;
 	float minVelocityVolume;
 	float maxVelocityVolume;
 	bool acceptsRetuning;